feat: framework of 'rationing' created
Its not actually rationing, but more to designed to represent the task of controlling the amount of consumer demand each good had.
The British public had accumulated a lot of savings during the war because there was nothing 'luxury' or 'consumerist' to spend it on.
Not having enough supply of goods would create demand-pull inflation, similar to the aftermath of Covid, so the UK government continued rationing as a form of price control to 'control' the demand.