feat: made multiple changes
I've removed all references to inflation. It now uses the concept of "Demand Pressure", which is basically inflation but different name. I've removed the weather system, but if in the future a weather system is implemented into the mod then that is something to use for this. Localization has been significantly changed, help the player understand what is happening and what they need to do to complete the journal entry.